+++ src/Motor.php
@@ -7,5 +7,5 @@ class Motor
     /**
      * Versión del núcleo del motor.
      */
-    public const VERSION = '0.0.1';
+    public const VERSION = '0.1.0';
 }